Restructure into domain-based architecture for scalability
- Create src/domains/ with invoices/, compliance/, notifications/, fiscal/ domains
  Each domain owns its business logic, types, and barrel index.ts
- Create src/types/fiscal.ts (shared UserFiscalProfile and related interfaces)
  and src/types/api.ts (ApiResponse<T>, PaginatedResponse<T> wrappers)
- Replace any[] with typed Prisma-shaped interfaces in fiscal-context.ts
- Convert lib/ domain files to re-export shims pointing to domains/ so
  all existing imports continue to work without changes
- Add @/domains/* and @/types/* path aliases to tsconfig.json
- Delete 4 unimported dead components (UrssafReportModal, UrssafReportsTable,
  SimulateurAutoEntrepreneur, TestTailwind)
