--- src/components/TestTailwind.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-import React from 'react';
-
-const TestTailwind = () => {
-  return (
-    <div className="bg-blue-500 text-white p-4 rounded-lg m-4">
-      This is a test component with Tailwind CSS.
-    </div>
-  );
-};
-
-export default TestTailwind;
