+++ src/domains/compliance/index.ts
@@ -1,0 +1,2 @@
+export * from './compliance';
+export * from './e-reporting/generate';
