+++ src/domains/fiscal/index.ts
@@ -1,0 +1,2 @@
+export * from './fiscal-context';
+export { default } from './fiscal-context';
