+++ src/domains/notifications/index.ts
@@ -1,0 +1,2 @@
+export * from './notification-service';
+export { default } from './notification-service';
