+++ src/lib/facturx/index.ts
@@ -1,3 +1,2 @@
-export * from "./invoice-to-cii";
-export * from "./facturx-generator";
-export * from "./facturx-validator";
+// Re-export from domain layer — import from @/domains/invoices/facturx instead
+export * from '@/domains/invoices/facturx';
